condensed non-segmenter functions into pre- and post-processing
