+++ ssts/utils/pre_processing.py
@@ -1,3 +1,5 @@
+import os
+
 import numpy as np
 from scipy import signal
 from matplotlib import pyplot, colors
@@ -6,15 +8,11 @@ from utils import config
 
 INFO = config.data_info
 
-LABEL_THRESH = 5  # each label must have more than this number of pixels
-
-ALPHA = 0.8  # transparency of labels in graphs
-NUM_BINS = 30  # number of bins in histograms
-
 NUM_COLS = 2  # number of cols in plots
 
 
-## Outlier Detection
+## Outlier detection and filtering methods
+
 def extract_outliers(data, data_type, threshold=2.5):
     """
     Finds outliers from data
@@ -213,3 +211,78 @@ def apply_frequency_removal(data, data_type, compression_percent=95):
     pyplot.show()
 
     return new_data
+
+## Data and file i/o methods
+
+def get_all_paths(path):
+    """
+    Gets list of data files in path
+    
+    Parameters
+    ----------
+        path : str
+            data directory
+    
+    Returns
+    ----------
+        fls : list
+            data files
+    """
+    fls = [os.path.join(path, f) for f in os.listdir(path)]
+    return fls
+
+
+def get_sample_count(path):
+    """
+    Gets number of data files in path
+    
+    Parameters
+    ----------
+        path : str
+            data directory
+        
+    Returns
+    ----------
+        int
+            number of data files
+    """
+    return len(get_all_paths(path))
+
+
+def get_path_from_index(index, path):
+    """
+    Gets specific data file from path
+    
+    Parameters
+    ----------
+        index : int
+            specific data file index
+        path : str
+            data directory
+    
+    Returns
+    ----------
+        str
+            data file
+    """
+    return get_all_paths(path)[index]
+
+
+def get_data(index, path):
+    """
+    Loads specific data file
+    
+    Parameters
+    ----------
+        index : int
+            specific data file index
+        path : str
+            data directory
+        
+    Returns
+    ----------
+       NumPy Array
+           data from file
+    """
+    fl = get_path_from_index(index, path)
+    return np.load(fl)
